refactor: move compute_active_finality_providers to power_dist_change.rs
Move compute_active_finality_providers and query_fps_by_total_active_sats functions from finality.rs to a new dedicated module power_dist_change.rs. This improves code organization by separating power distribution logic from general finality operations.

Changes:
- Create new power_dist_change.rs module
- Move compute_active_finality_providers function
- Move query_fps_by_total_active_sats helper function
- Move QUERY_LIMIT and JAIL_FOREVER constants
- Update imports in contract.rs and finality.rs
- Add module declaration to lib.rs
