+++ contracts/btc-finality/src/finality.rs
@@ -1,25 +1,24 @@
-use crate::contract::encode_smart_query;
 use crate::error::{ContractError, FinalitySigError};
 use crate::msg::{MsgAddFinalitySig, MsgCommitPubRand};
+use crate::power_dist_change::JAIL_FOREVER;
 use crate::state::config::{ADMIN, CONFIG};
 use crate::state::finality::{
-    collect_accumulated_voting_weights, get_fp_power, get_last_signed_height,
-    get_power_table_at_height, set_voting_power_table, ACCUMULATED_VOTING_WEIGHTS, BLOCKS,
+    collect_accumulated_voting_weights, get_fp_power,
+    get_power_table_at_height, ACCUMULATED_VOTING_WEIGHTS, BLOCKS,
     EVIDENCES, FP_BLOCK_SIGNER, FP_START_HEIGHT, JAIL, NEXT_HEIGHT, SIGNATURES,
 };
 use crate::state::public_randomness::{
-    get_last_finalized_height, get_last_pub_rand_commit,
-    get_timestamped_pub_rand_commit_for_height, has_timestamped_pub_rand_commit_for_height,
+    get_last_pub_rand_commit,
+    get_timestamped_pub_rand_commit_for_height,
     PUB_RAND_COMMITS, PUB_RAND_VALUES,
 };
 use babylon_apis::btc_staking_api::FinalityProvider;
 use babylon_apis::finality_api::{Evidence, IndexedBlock, PubRandCommit};
 use babylon_apis::to_canonical_addr;
 use babylon_contract::msg::contract::{ExecuteMsg, RewardInfo};
-use btc_staking::msg::{FinalityProviderInfo, FinalityProvidersByTotalActiveSatsResponse};
 use cosmwasm_std::Order::Ascending;
 use cosmwasm_std::{
-    coins, to_json_binary, Addr, DepsMut, Env, Event, MessageInfo, QuerierWrapper, Response,
+    coins, to_json_binary, DepsMut, Env, Event, MessageInfo, Response,
     StdResult, Storage, Uint128, WasmMsg,
 };
 use std::cmp::max;
@@ -31,9 +30,7 @@ use std::collections::{HashMap, HashSet};
 // or performance degradation caused by excessive future commitments.
 const MAX_PUB_RAND_COMMIT_OFFSET: u64 = 160_000;
 
-const QUERY_LIMIT: Option<u32> = Some(30);
 
-pub const JAIL_FOREVER: u64 = 0;
 
 pub fn handle_public_randomness_commit(
     deps: DepsMut,
@@ -532,108 +529,7 @@ fn finalize_block(
     Ok(ev)
 }
 
-/// Sorts all finality providers, counts the total voting power of top finality providers, and records them
-/// in the contract state.
-pub fn compute_active_finality_providers(
-    deps: &mut DepsMut,
-    env: &Env,
-    max_active_fps: usize,
-) -> Result<(), ContractError> {
-    let cfg = CONFIG.load(deps.storage)?;
-    // Get last finalized height (for timestamped public randomness checks)
-    let last_finalized_height = get_last_finalized_height(&deps.as_ref())?;
 
-    // Get all finality providers from the staking contract, filtered
-    let mut batch = query_fps_by_total_active_sats(&cfg.staking, &deps.querier, None, QUERY_LIMIT)?;
-
-    let mut fp_power_table = HashMap::new();
-    let mut total_power: u64 = 0;
-    while !batch.is_empty() && fp_power_table.len() < max_active_fps {
-        let last = batch.last().cloned();
-
-        let (filtered, running_total): (Vec<_>, Vec<_>) = batch
-            .into_iter()
-            .filter(|fp| {
-                // Filter out FPs with no active sats
-                if fp.total_active_sats == 0 {
-                    return false;
-                }
-                // Filter out slashed FPs
-                if fp.slashed {
-                    return false;
-                }
-                // Filter out FPs that are jailed.
-                // Error (shouldn't happen) is being mapped to "jailed forever"
-                if JAIL
-                    .may_load(deps.storage, &fp.btc_pk_hex)
-                    .unwrap_or(Some(JAIL_FOREVER))
-                    .is_some()
-                {
-                    return false;
-                }
-                // Filter out FPs that don't have timestamped public randomness
-                if !has_timestamped_pub_rand_commit_for_height(
-                    &deps.as_ref(),
-                    &fp.btc_pk_hex,
-                    env.block.height,
-                    Some(last_finalized_height),
-                ) {
-                    return false;
-                }
-
-                true
-            })
-            .scan(total_power, |acc, fp| {
-                *acc += fp.total_active_sats;
-                Some((fp, *acc))
-            })
-            .unzip();
-
-        // Add the filtered finality providers to the power table
-        for fp in filtered {
-            fp_power_table.insert(fp.btc_pk_hex, fp.total_active_sats);
-        }
-        // Update the total power
-        total_power = running_total.last().copied().unwrap_or_default();
-
-        // and get the next page
-        batch = query_fps_by_total_active_sats(&cfg.staking, &deps.querier, last, QUERY_LIMIT)?;
-    }
-
-    // Online FPs verification
-    // Store starting heights of fps entering the active set
-    let old_power_table = get_power_table_at_height(deps.storage, env.block.height - 1)?;
-    let old_fps = old_power_table.keys().collect();
-    let cur_fps: HashSet<_> = fp_power_table.keys().collect();
-    let new_fps = cur_fps.difference(&old_fps);
-    for fp in new_fps {
-        // Active since the next block. Only save if not already set
-        FP_START_HEIGHT.update(deps.storage, fp, |h| match h {
-            Some(h) => Ok::<_, ContractError>(h),
-            None => Ok(env.block.height + 1),
-        })?;
-    }
-
-    // Save the new set of active finality providers
-    set_voting_power_table(deps.storage, env.block.height, fp_power_table)?;
-
-    Ok(())
-}
-
-/// Queries the BTC staking contract for finality providers ordered by total active sats.
-pub fn query_fps_by_total_active_sats(
-    staking_addr: &Addr,
-    querier: &QuerierWrapper,
-    start_after: Option<FinalityProviderInfo>,
-    limit: Option<u32>,
-) -> StdResult<Vec<FinalityProviderInfo>> {
-    let query = encode_smart_query(
-        staking_addr,
-        &btc_staking::msg::QueryMsg::FinalityProvidersByTotalActiveSats { start_after, limit },
-    )?;
-    let res: FinalityProvidersByTotalActiveSatsResponse = querier.query(&query)?;
-    Ok(res.fps)
-}
 
 /// Handles finality provider reward distribution based on accumulated voting weights.
 ///
